spinnaker_camera_driver: expose AE ROI parameters
This is highly useful when using fisheye lenses, which illuminate only
a circle in the center of the image. The AE gets confused by the black
regions around it and overexposes the image.

This also exposes the "AutoExposureLightingMode" parameter, which allows
the user to choose a lighting preset (front/back/normal).